+++ webpack/webpack.common.config.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+const config = {
+    entry: './src/js/index.js',
+    output: {
+        path: path.resolve(__dirname, '../dist'),
+        filename: 'bundle.js',
+    },
+}
+
+module.exports = config
